web/requirements/default.txt
Flask>=0.10.1,<0.11
arrow
gevent>=1.1.1
python-decouple
pymongo>=3.0.3
Flask-Script
Flask-Moment
#Flask-Bootstrap-3.3.5.7 - bootstrap 3.3.5 - jquery 1.11.3
Flask-Bootstrap
Flask-BasicAuth
Flask-Mail
redis
psutil
Flask-Assets
jsmin
cssmin